refactor: Dokan REST base controllers and docs
