+++ includes/Abstracts/DokanRESTAdminController.php
@@ -2,30 +2,15 @@
 
 namespace WeDevs\Dokan\Abstracts;
 
+use WeDevs\Dokan\REST\DokanBaseAdminController;
+
 /**
 * Admin Dashboard
 *
 * @since 2.8.0
+* @deprecated DOKAN_SINCE Use \WeDevs\Dokan\REST\DokanBaseAdminController instead.
 *
 * @package dokan
 */
-abstract class DokanRESTAdminController extends DokanRESTBaseController {
-
-    /**
-     * Endpoint namespace.
-     *
-     * @var string
-     */
-    protected $namespace = 'dokan/v1/admin';
-
-    /**
-     * Check if user has admin permission.
-     *
-     * @since 2.8.0
-     *
-     * @return bool
-     */
-    public function check_permission() {
-        return current_user_can( 'manage_woocommerce' );
-    }
+abstract class DokanRESTAdminController extends DokanBaseAdminController {
 }
